refactor: Simplify storages OpenAPI implementation
- Remove unnecessary openapi-handlers.ts wrapper file
- Consolidate openapi-routes.ts into route.ts for cleaner structure
- Export OpenAPI route definitions alongside traditional Hono router
- Update api/index.ts imports to use consolidated route.ts
- Maintain backwards compatibility with existing route handlers

This reduces file complexity while keeping all storage-related routes
in a single location, making the codebase easier to maintain.
