+++ src/api/index.ts
@@ -9,8 +9,7 @@ import { setupGetHandler } from './v1/setup/get';
 import { setupPostHandler } from './v1/setup/post';
 import { createSheetRoute } from './v1/sheets/route';
 import { sheetsPostHandler } from './v1/sheets/post';
-import storagesRouter from './v1/storages/route';
-import { uploadFileRoute, deleteFileRoute } from './v1/storages/openapi-routes';
+import storagesRouter, { uploadFileRoute, deleteFileRoute } from './v1/storages/route';
 import storagesPostHandler from './v1/storages/post';
 import storagesDeleteHandler from './v1/storages/delete';
 import authRouter from './v1/auth';
